pyproject.toml
[project]
name = "lsp-cli"
version = "0.1.1"
description = "A powerful command-line interface for the Language Server Agent Protocol (LSAP)"
readme = "README.md"
authors = [{ name = "observerw", email = "wozluohd@gmail.com" }]
requires-python = ">=3.13"
keywords = ["Language Server Protocol", "LSP", "CLI", "Developer Tools", "LSAP"]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "anyio>=4.12.0",
    "asyncer>=0.0.12",
    "attrs>=25.4.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "platformdirs>=4.5.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "typer>=0.20.1",
    "uvicorn>=0.40.0",
    "litestar>=2.19.0",
    "xxhash>=3.6.0",
    "tenacity>=9.1.2",
    "lsp-client>=0.3.0",
    "lsap-sdk>=0.1.0",
]

[project.scripts]
lsp = "lsp_cli.__main__:run"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.pytest.ini_options]
norecursedirs = ["references"]
